extract common Flask/Blueprint API to Scaffold base class
